Arreglar para entrega

+++ src/main/java/com/pentabyte/projects/sorteador/dto/response/initial/GlobalDTO.java
@@ -1,16 +1,17 @@
 package com.pentabyte.projects.sorteador.dto.response.initial;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record GlobalDTO(Integer totales,
-                        Integer planificadas,
-                        Integer borrador,
-                        Integer pendientes,
-                        Integer noPendientes,
-                        Integer confirmados,
-                        Integer noConfirmados
+public record GlobalDTO(Long totales,
+                        Long planificadas,
+                        Long borrador,
+                        Long pendientes,
+                        Long noPendientes,
+                        Long confirmados,
+                        Long noConfirmados
 ) {
 
 }
